initial config builder + DI support

+++ TagsCloudContainerCore/Models/Tag.cs
@@ -4,7 +4,7 @@ namespace TagsCloudContainerCore.Models;
 
 public class Tag
 {
-    public int FontSize;
+    public float FontSize;
     public required Color Color;
     public required Rectangle BBox;
     public required string Text;
